+++ src/Symfony/Component/Ldap/Entry.php
@@ -13,16 +13,23 @@ namespace Symfony\Component\Ldap;
 
 /**
  * @author Charles Sarrazin <charles@sarraz.in>
+ * @author Karl Shea <karl@karlshea.com>
  */
 class Entry
 {
     private $dn;
     private $attributes;
+    private $lowerMap;
 
     public function __construct(string $dn, array $attributes = [])
     {
         $this->dn = $dn;
-        $this->attributes = $attributes;
+        $this->lowerMap = [];
+
+        foreach ($attributes as $key => $attribute)
+        {
+            $this->setAttribute($key, $attribute);
+        }
     }
 
     /**
@@ -39,12 +46,19 @@ class Entry
      * Returns whether an attribute exists.
      *
      * @param string $name The name of the attribute
+     * @param bool $caseSensitive Whether the check should be case-sensitive
      *
      * @return bool
      */
-    public function hasAttribute(string $name)
+    public function hasAttribute(string $name, $caseSensitive = true)
     {
-        return isset($this->attributes[$name]);
+        $attributeKey = $this->getAttributeKey($name, $caseSensitive);
+
+        if (!$attributeKey) {
+            return false;
+        }
+
+        return isset($this->attributes[$attributeKey]);
     }
 
     /**
@@ -54,12 +68,19 @@ class Entry
      * this value is returned as an array.
      *
      * @param string $name The name of the attribute
+     * @param bool $caseSensitive Whether the attribute name is case-sensitive
      *
      * @return array|null
      */
-    public function getAttribute(string $name)
+    public function getAttribute(string $name, $caseSensitive = true)
     {
-        return isset($this->attributes[$name]) ? $this->attributes[$name] : null;
+        $attributeKey = $this->getAttributeKey($name, $caseSensitive);
+
+        if (!$attributeKey) {
+            return null;
+        }
+
+        return isset($this->attributes[$attributeKey]) ? $this->attributes[$attributeKey] : null;
     }
 
     /**
@@ -78,6 +99,7 @@ class Entry
     public function setAttribute(string $name, array $value)
     {
         $this->attributes[$name] = $value;
+        $this->lowerMap[strtolower($name)] = $name;
     }
 
     /**
@@ -86,5 +108,23 @@ class Entry
     public function removeAttribute(string $name)
     {
         unset($this->attributes[$name]);
+        unset($this->lowerMap[strtolower($name)]);
+    }
+
+    /**
+     * Get the attribute key.
+     *
+     * @param string $name The attribute name
+     * @param bool $caseSensitive Whether the attribute name is case-sensitive
+     *
+     * @return string|null
+     */
+    private function getAttributeKey(string $name, $caseSensitive = true)
+    {
+        if ($caseSensitive) {
+            return $name;
+        }
+
+        return isset($this->lowerMap[strtolower($name)]) ? $this->lowerMap[strtolower($name)] : null;
     }
 }